refactor: replace ExampleView with HowToUseView and update routing
